Chef-15 DSL changes
Move the data_query, powershell and registry_helper DSLs up to universal.

Move the `tagged?` helper into data_query for organizational purposes.
